backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
spacy
sentence-transformers
pdfplumber
python-multipart
requests
passlib[bcrypt]
python-jose[cryptography]
email-validator
